Remove old benchmark infra
