--- NewBenchmarks/Models/ResNetCIFAR10.swift
@@ -1,72 +1,0 @@
-// Copyright 2020 The TensorFlow Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import Benchmark
-import Datasets
-import ImageClassificationModels
-import TensorFlow
-
-let ResNetCIFAR10 = BenchmarkSuite(name: "ResNetCIFAR10") { suite in
-  suite.benchmark("inference", settings: BatchSize(128), WarmupIterations(1)) {
-    state in
-    if state.settings.synthetic! {
-      try runImageClassificationInference(
-        model: ResNet56.self, dataset: SyntheticCIFAR10.self, state: &state)
-    } else {
-      try runImageClassificationInference(
-        model: ResNet56.self, dataset: CIFAR10<SystemRandomNumberGenerator>.self, state: &state)
-    }
-  }
-
-  suite.benchmark("training", settings: BatchSize(128), WarmupIterations(1)) {
-    state in
-    if state.settings.synthetic! {
-      try runImageClassificationTraining(
-        model: ResNet56.self, dataset: SyntheticCIFAR10.self, state: &state)
-    } else {
-      try runImageClassificationTraining(
-        model: ResNet56.self, dataset: CIFAR10<SystemRandomNumberGenerator>.self, state: &state)
-    }
-
-  }
-}
-
-struct ResNet56: Layer {
-  var model: ResNet
-
-  init() {
-    model = ResNet(classCount: 10, depth: .resNet56, downsamplingInFirstStage: false)
-  }
-
-  @differentiable
-  public func callAsFunction(_ input: Tensor<Float>) -> Tensor<Float> {
-    return model(input)
-  }
-}
-
-extension ResNet56: ImageClassificationModel {
-  static var preferredInputDimensions: [Int] { [32, 32, 3] }
-  static var outputLabels: Int { 10 }
-}
-
-final class SyntheticCIFAR10: SyntheticImageDataset<SystemRandomNumberGenerator>,
-  ImageClassificationData
-{
-  public init(batchSize: Int, on device: Device = Device.default) {
-    super.init(
-      batchSize: batchSize, labels: ResNet56.outputLabels,
-      dimensions: ResNet56.preferredInputDimensions, entropy: SystemRandomNumberGenerator(),
-      device: device)
-  }
-}
